# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indo-German Mobility Corridor Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visa-cases:
    get:
      operationId: listVisaCases
      tags: [visa]
      summary: List visa cases
      responses:
        "200":
          description: Visa cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisaCase"
    post:
      operationId: createVisaCase
      tags: [visa]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisaCaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisaCase"
  /visa-cases/{id}:
    get:
      operationId: getVisaCase
      tags: [visa]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visa case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisaCase"
    patch:
      operationId: updateVisaCase
      tags: [visa]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisaCaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisaCase"
  /visa-cases/{id}/checklist:
    get:
      operationId: getVisaChecklist
      tags: [visa]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checklist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisaChecklistItem"
  /visa-cases/{id}/checklist/{itemId}:
    patch:
      operationId: updateVisaChecklistItem
      tags: [visa]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisaChecklistItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisaChecklistItem"

  # ── INTERVIEWS ───────────────────────────────────────────────────────────────
...
  /welfare-tickets:
    get:
      operationId: listWelfareTickets
      tags: [welfare]
      summary: List welfare/grievance tickets
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WelfareTicket"
    post:
      operationId: createWelfareTicket
      tags: [welfare]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WelfareTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelfareTicket"
  /welfare-tickets/{id}:
    get:
      operationId: getWelfareTicket
      tags: [welfare]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelfareTicket"
    patch:
      operationId: updateWelfareTicket
      tags: [welfare]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WelfareTicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelfareTicket"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    VisaCase:
      type: object
      required: [id, candidateId, status, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        pathwayType:
          type: ["string", "null"]
        status:
          type: string
          enum: [not_started, checklist_in_progress, documents_ready, appointment_pending, submitted, decision_pending, approved, rejected]
        checklistProgress:
          type: integer
        checklistTotal:
          type: integer
        travelDate:
          type: ["string", "null"]
        arrivalConfirmed:
          type: boolean
        createdAt:
          type: string
    VisaCaseInput:
      type: object
      required: [candidateId]
      properties:
        candidateId:
          type: integer
        pathwayType:
          type: string
    VisaCaseUpdate:
      type: object
      properties:
        status:
          type: string
        travelDate:
          type: string
        arrivalConfirmed:
          type: boolean
    VisaChecklistItem:
      type: object
      required: [id, visaCaseId, item, completed]
      properties:
        id:
          type: integer
        visaCaseId:
          type: integer
        item:
          type: string
        description:
          type: ["string", "null"]
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        required:
          type: boolean
    VisaChecklistItemUpdate:
      type: object
      properties:
        completed:
          type: boolean

    # ── INTERVIEW ──
...
    WelfareTicket:
      type: object
      required: [id, candidateId, category, title, status, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        category:
          type: string
          enum: [accommodation, exploitation, fraud, health, legal, general, employer_grievance]
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        assignedTo:
          type: ["string", "null"]
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
    WelfareTicketInput:
      type: object
      required: [category, title, description]
      properties:
        category:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
    WelfareTicketUpdate:
      type: object
      properties:
        status:
          type: string
        assignedTo:
          type: string
        resolution:
          type: string
        priority:
          type: string

    # ── NOTIFICATION ──
...